test(zcode): split hook and MCP tests under 500-line limit
Move TestZCodeHookPayload and TestZCodeHookResponses into a dedicated
tests/test_zcode_hooks.py and the MCP PreToolUse coverage regression into
tests/test_zcode_mcp.py so each ZCode test file stays within the 500-line
project limit. No test behavior changes.
